feat: Add EditorGroupsOpenListener and update plugin.xml with required listener
Adds EditorGroupsOpenListener to handle fileOpenedSync event for creating editor group panels.
Updates plugin.xml to include the required listener for the new functionality.
