+++ src/main/java/krasa/editorGroups/listeners/EditorGroupsOpenListener.kt
@@ -1,4 +1,4 @@
-package krasa.editorGroups
+package krasa.editorGroups.listeners
 
 import com.intellij.ide.ui.UISettings
 import com.intellij.ide.ui.UISettingsListener
@@ -7,27 +7,34 @@ import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.fileEditor.FileEditor
 import com.intellij.openapi.fileEditor.FileEditorManager
-import com.intellij.openapi.fileEditor.FileEditorManagerEvent
-import com.intellij.openapi.fileEditor.FileEditorManagerListener
+import com.intellij.openapi.fileEditor.FileOpenedSyncListener
+import com.intellij.openapi.fileEditor.ex.FileEditorWithProvider
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
+import krasa.editorGroups.EditorGroupManager
+import krasa.editorGroups.EditorGroupPanel
+import krasa.editorGroups.SwitchRequest
 import krasa.editorGroups.support.unwrapPreview
 import javax.swing.SwingConstants
 
-class EditorGroupsStartup : FileEditorManagerListener {
-  override fun fileOpened(manager: FileEditorManager, file: VirtualFile) {
+class EditorGroupsOpenListener : FileOpenedSyncListener {
+  override fun fileOpenedSync(
+    manager: FileEditorManager,
+    file: VirtualFile,
+    editorsWithProviders: List<FileEditorWithProvider>
+  ) {
     val project = manager.project
     thisLogger().debug(">fileOpenedSync [$file]")
 
     val fileToOpen = unwrapPreview(file) ?: return
-    val editorGroupManager = EditorGroupManager.getInstance(project)
+    val editorGroupManager = EditorGroupManager.Companion.getInstance(project)
 
     val switchRequest = editorGroupManager.getAndClearSwitchingRequest(fileToOpen)
     val editors = manager.getEditors(fileToOpen)
 
     // Create editor group panel if it doesn't exist'
     for (fileEditor in editors) {
-      if (fileEditor.getUserData(EditorGroupPanel.EDITOR_PANEL) != null) continue
+      if (fileEditor.getUserData(EditorGroupPanel.Companion.EDITOR_PANEL) != null) continue
 
       val start = System.currentTimeMillis()
 
@@ -42,32 +49,6 @@ class EditorGroupsStartup : FileEditorManagerListener {
       thisLogger().debug("<fileOpenedSync EditorGroupPanel created, file=$fileToOpen in ${System.currentTimeMillis() - start}ms, fileEditor=$fileEditor")
     }
   }
-
-  /** When a tab is selected, refresh the editor group panel. */
-  override fun selectionChanged(event: FileEditorManagerEvent) {
-    val project = event.manager.project
-    thisLogger().debug("selectionChanged $event")
-    val fileEditor = event.newEditor
-    if (fileEditor == null) return
-
-    val panel = fileEditor.getUserData(EditorGroupPanel.EDITOR_PANEL)
-    if (panel == null) return
-
-    val instance = EditorGroupManager.getInstance(project)
-    val switchRequest = instance.getAndClearSwitchingRequest(panel.file)
-
-    if (switchRequest != null) {
-      val switchingGroup = switchRequest.group
-      val scrollOffset = switchRequest.myScrollOffset
-
-      // Refresh panel
-      panel.refreshOnSelectionChanged(false, switchingGroup, scrollOffset)
-    } else {
-      panel.refreshPane(false, null)
-    }
-  }
-
-  override fun fileClosed(source: FileEditorManager, file: VirtualFile) = thisLogger().debug("fileClosed [$file]")
 
   /**
    * Creates the editor group panel
